submit B2.cpp to 'B - ISU' (abc078) [C++14 (GCC 5.4.1)]

+++ 2018/1116_ABC078/B2.cpp
@@ -40,7 +40,7 @@ typedef long long ll;
 
 ll X, Y, Z;
 
-int sum(int n)
+ll sum(int n)
 {
   return n * Y + (n + 1) * Z;
 }
